+++ lib/visual_snapshots/run-android-dextrose.sh
@@ -2,7 +2,7 @@
 npx appium &
 APPIUM_PID=$!
 LOCAL=true npx fructose-tunnel
-emulator @NEXUS_5X_API_22 -no-boot-anim &
+emulator @Nexus_5X_API_25 -no-boot-anim &
 EMU_PID=$!
 adb wait-for-device
 adb reverse tcp:8081 tcp:8081
@@ -13,7 +13,7 @@ npx lerna run dextrose-clean
 npx lerna run dextrose-stories --since
 npx compile-tests -d fructose
 npx rnstl --searchDir ./packages --pattern './*/*.dextrose.tmp.js' --outputFile ./fructose/components.js
-react-native start --root fructose &
+npx react-native start --root fructose --projectRoots $(pwd)/fructose,$(pwd) &
 react-native run-android --variant=debug --no-packager
 LOGLEVEL=verbose npx dextrose --config ./dextrose/dextrose.android.js --timeout 600000  --snapshotWait 2000
 npx lerna run dextrose-clean