+++ packages/article-label/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.6.2"></a>
+## [0.6.2](https://github.com/newsuk/times-components/compare/@times-components/article-label@0.6.1...@times-components/article-label@0.6.2) (2017-12-14)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-label
+
 <a name="0.6.1"></a>
 ## [0.6.1](https://github.com/newsuk/times-components/compare/@times-components/article-label@0.6.0...@times-components/article-label@0.6.1) (2017-12-13)